drivers: usb: device: sam0:  Fix EP write buf failed
The sam0 driver checks the status of the endpoint buffer but does
not handle the stuation where the buffer is not ready.  This issue becomes
a problem when large volumes of data and written and the hardware
mechanisms controlling the buffer are busy.  This results in spurious
"Failed to write endpoint buffer" warning messages to the console.
This commit fixes the issue by inserting a short busy
wait, wtih timeout, for the buffer to become available.

Signed-off-by: Steven Slupsky <sslupsky@gmail.com>